requirements.txt
pyTelegramBotAPI
python-dotenv
requests
CurrencyConverter
